+++ CHANGELOG.md
@@ -26,6 +26,9 @@ Breaking changes
 * @iov/ethereum: The `options` parameter of `ethereumConnector` does not allow
   strings anymore. Use `{ wsUrl: myWebsocketUrl }` instead.
 * @iov/keycontrol: Remove `HdPaths.metamaskHdKeyTree`. Use `HdPaths.ethereum` instead.
+* @iov/keycontrol: `Keyring.getWallet` and `Keyring.getWallets` return the
+  immutable type ReadonlyWallet now. New functions to mutate wallets are added:
+  `Keyring.setWalletLabel`, `.createIdentity`, `.setIdentityLabel` are added.
 * @iov/ledger-bns: Package removed from this monorepo and now available at
   https://github.com/iov-one/iov-ledger-bns